+++ lib/src/components/member/member_list_button.dart
@@ -49,8 +49,8 @@ class _ZegoMemberListButtonState extends State<ZegoMemberListButton> {
 
   @override
   Widget build(BuildContext context) {
-    final containerSize = widget.buttonSize ?? Size(96.r, 96.r);
-    final sizeBoxSize = widget.iconSize ?? Size(56.r, 56.r);
+    final containerSize = widget.buttonSize ?? Size(96.zR, 96.zR);
+    final sizeBoxSize = widget.iconSize ?? Size(56.zR, 56.zR);
 
     return GestureDetector(
       onTap: () {